+++ example/lib/src/version/notifiers/version_notifier.dart
@@ -1,14 +1,13 @@
 import 'package:minimal_mvn/minimal_mvn.dart';
 
 import '../utils/version.dart';
-import '../views/ui_states/version_ui_state.dart';
 
-class VersionNotifier extends MMNotifier<VersionUIState> {
-  VersionNotifier() : super(const VersionUIState());
+class VersionNotifier extends MMNotifier<String> {
+  VersionNotifier() : super('');
 
   Future<void> loadVersion() async {
     final version = await getPackageVersion();
-    notify(VersionUIState(version: version));
+    notify(version);
   }
 }