+++ lib/minimal_mvn.dart
@@ -6,7 +6,6 @@ import 'minimal_mvn.dart';
 /// applications using the MVN (Model-View-Notifier) pattern.
 ///
 /// The main components are:
-/// * [MMState] - Minimal state for all immutable states
 /// * [MMNotifier] - Minimal notifier for all notifiers that manage state
 /// * [MMManager] - Minimal manager for accessing notifiers
 /// * [MMLocator] - Minimal locator for accessing repositories and services
@@ -14,4 +13,3 @@ import 'minimal_mvn.dart';
 export 'src/mm_locator.dart';
 export 'src/mm_manager.dart';
 export 'src/mm_notifier.dart';
-export 'src/mm_state.dart';