+++ lib/src/mm_notifier.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/foundation.dart';
-import 'mm_state.dart';
 
 /// Callback invoked when a notifier has no more subscribers
 typedef OnUnsubscribedCallback = void Function();
@@ -17,7 +16,7 @@ typedef OnUnsubscribedCallback = void Function();
 ///   void increment() => notify(state.copyWith(count: state.count + 1));
 /// }
 /// ```
-abstract class MMNotifier<T extends MMState> extends ChangeNotifier {
+abstract class MMNotifier<T> extends ChangeNotifier {
   /// Creates a minimal notifier with the given initial state
   MMNotifier(this._state);
 
@@ -54,7 +53,7 @@ abstract class MMNotifier<T extends MMState> extends ChangeNotifier {
   /// when it changes
   ///
   /// This is useful for optimizing rebuilds in the UI. The returned
-  /// [ValueNotifier] will only notify its listeners when the selected value
+  /// [ValueListenable] will only notify its listeners when the selected value
   /// actually changes
   ///
   /// Example:
@@ -69,7 +68,7 @@ abstract class MMNotifier<T extends MMState> extends ChangeNotifier {
   ///   ),
   /// );
   /// ```
-  ValueNotifier<S> select<S>(final S Function(T state) selector) {
+  ValueListenable<S> select<S>(final S Function(T state) selector) {
     final notifier = _MMSelector(() => selector(_state));
     // the moment selector is subscribed, start listening to the notifier
     // for its changes
@@ -84,15 +83,19 @@ abstract class MMNotifier<T extends MMState> extends ChangeNotifier {
   /// Mutates the notifier's state and notifies its listeners
   ///
   /// This is the preferred way to update state as it ensures all listeners
-  /// are notified
+  /// are notified. In case the new state is the same as the current state,
+  /// the notification is skipped to avoid unnecessary rebuilds
   ///
   /// Example:
   /// ```dart
   /// void increment() => notify(state.copyWith(value: state.value + 1));
   /// ```
+  @protected
   void notify(final T value) {
-    _state = value;
-    notifyListeners();
+    if (_state != value) {
+      _state = value;
+      notifyListeners();
+    }
   }
 
   /// Callback that will be invoked when this notifier has no more
@@ -131,10 +134,7 @@ class _MMSelector<T> extends ValueNotifier<T> {
   VoidCallback? onRemoveListener;
 
   void notify() {
-    final current = _getValue();
-    if (current != value) {
-      value = current;
-    }
+    value = _getValue();
   }
 
   @override