+++ test/src/mm_notifier_test.dart
@@ -1,5 +1,6 @@
 // Let's make the tests more readable and explicit
 // ignore_for_file: cascade_invocations
+import 'package:flutter/foundation.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'utils/test_common.dart';
@@ -57,22 +58,41 @@ void main() {
       selected.addListener(selectionL);
 
       notifier.increment();
+      expect(selected.value, '');
       expect(notifier.state.value, 1);
       expect(notifier.state.text, '');
       expect(stateNotified, 1);
       expect(selectionNotified, 0);
 
       notifier.increment();
+      expect(selected.value, '');
       expect(notifier.state.value, 2);
       expect(notifier.state.text, '');
       expect(stateNotified, 2);
       expect(selectionNotified, 0);
 
       notifier.append('mutated');
+      expect(selected.value, 'mutated');
       expect(notifier.state.value, 2);
       expect(notifier.state.text, 'mutated');
       expect(stateNotified, 3);
       expect(selectionNotified, 1);
+    });
+
+    test('selection cannot be mutated', () {
+      var selectionNotified = 0;
+      void selectionL() => selectionNotified++;
+
+      final notifier = TNotifier();
+
+      final selected = notifier.select((final state) => state.text);
+      selected.addListener(selectionL);
+
+      // ValueListenable is read-only, so we can't mutate it
+      // ignore: unnecessary_type_check
+      expect(selected is ValueListenable, true);
+      notifier.append('parent update');
+      expect(selected.value, 'parent update');
     });
 
     test('tracks disposed state', () {